requirements.txt
fastapi
uvicorn
boto3
pytest
httpx
python-dotenv